src/requirements.txt
Flask==2.2.2
pandas==1.5.1
Pillow==9.2.0
py==1.10.0
pycodestyle==2.7.0
pydocstyle==6.0.0
pytest==6.2.3
pytest-cov==2.11.1
python-docx==0.8.11
requests==2.28.1
requests-mock==1.9.2
setuptools-scm==6.0.1
wheel==0.36.2